requirements.txt
# Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Anomaly Detection
pyod>=1.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Streamlit
streamlit>=1.28.0

# Multi-Agent
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.2.0
openai>=1.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
